+++ internal/acceptance_tests/blocks/logging_ftp_compute_format.tf
@@ -1,0 +1,12 @@
+resource "fastly_service_logging_ftp" "test" {
+  service_id = fastly_service_compute.test.id
+  version    = {{.SERVICE_VERSION}}
+  name       = "{{.LOGGING_FTP_NAME}}"
+  address    = "ftp.example.com"
+  path       = "/"
+  authentication = {
+    user     = "test-user"
+    password = "test-password"
+  }
+  format = "%h %l %u %t \"%r\" %>s %b"
+}