+++ internal/acceptance_tests/blocks/logging_ftp_nested_multi.tf
@@ -1,0 +1,19 @@
+logging_ftp {
+  name    = "{{.LOGGING_FTP_NAME_1}}"
+  address = "ftp.example.com"
+  path    = "/"
+  authentication = {
+    user     = "test-user"
+    password = "test-password"
+  }
+}
+
+logging_ftp {
+  name    = "{{.LOGGING_FTP_NAME_2}}"
+  address = "ftp.example.com"
+  path    = "/"
+  authentication = {
+    user     = "test-user"
+    password = "test-password"
+  }
+}