+++ internal/acceptance_tests/blocks/logging_ftp_nested_placement_none.tf
@@ -1,0 +1,10 @@
+logging_ftp {
+  name    = "{{.LOGGING_FTP_NAME}}"
+  address = "ftp.example.com"
+  path    = "/"
+  authentication = {
+    user     = "test-user"
+    password = "test-password"
+  }
+  placement = "none"
+}