+++ internal/provider/provider.go
@@ -95,6 +95,7 @@ import (
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingdatadog"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingdigitalocean"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingelasticsearch"
+	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingftp"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/logginggcs"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/logginggrafanacloudlogs"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/logginghttps"
@@ -241,6 +242,7 @@ func (p *fastlyProvider) Resources(_ context.Context) []func() resource.Resource
 		loggingdatadog.NewResource,
 		loggingdigitalocean.NewResource,
 		loggingelasticsearch.NewResource,
+		loggingftp.NewResource,
 		logginggcs.NewResource,
 		logginggrafanacloudlogs.NewResource,
 		logginghttps.NewResource,
@@ -383,6 +385,7 @@ func (p *fastlyProvider) ListResources(_ context.Context) []func() list.ListReso
 		loggingdatadog.NewListResource,
 		loggingdigitalocean.NewListResource,
 		loggingelasticsearch.NewListResource,
+		loggingftp.NewListResource,
 		logginggcs.NewListResource,
 		logginggrafanacloudlogs.NewListResource,
 		logginghttps.NewListResource,